// backend/src/constants/food.constants.d.ts
export declare enum MealType {
    BREAKFAST = "BREAKFAST",
    LUNCH = "LUNCH",
    SNACK = "SNACK",
    DINNER = "DINNER"
}
export declare enum RecipeCategory {
    BREAKFAST = "BREAKFAST",
    LUNCH = "LUNCH",
    SNACK = "SNACK",
    DINNER = "DINNER",
    DESERT = "DESERT",
    SMOOTHIE = "SMOOTHIE"
}
export declare enum DietType {
    NO_RESTRICTION = "NO_RESTRICTION",
    VEGETARIAN = "VEGETARIAN",
    VEGAN = "VEGAN",
    KETO = "KETO",
    PALEO = "PALEO"
}
export declare enum Difficulty {
    EASY = "EASY",
    MEDIUM = "MEDIUM",
    HARD = "HARD"
}
//# sourceMappingURL=food.constants.d.ts.map